ie: fix compilation error on 64 bit
src/ie.c:33: error: conflicting types for 'dect_ie_alloc'
include/dect/ie.h:49: error: previous declaration of 'dect_ie_alloc' was here
src/ie.c:44: error: conflicting types for 'dect_ie_alloc'
include/dect/ie.h:49: error: previous declaration of 'dect_ie_alloc' was here

+++ src/ie.c
@@ -30,8 +30,7 @@
  * Information Elements
  */
 
-struct dect_ie_common *dect_ie_alloc(const struct dect_handle *dh,
-				     unsigned int size)
+struct dect_ie_common *dect_ie_alloc(const struct dect_handle *dh, size_t size)
 {
 	struct dect_ie_common *ie;
